fix: change reset-password & update-password DTOS
