+++ src/api/user/dtos/update-password.dto.ts
@@ -1,18 +1,8 @@
 import { ApiProperty } from "@nestjs/swagger";
-import { IsEmail, IsNotEmpty, IsString, Matches } from "class-validator";
+import { IsNotEmpty, IsString, Matches } from "class-validator";
 import { SameAs } from "../../../common/decorators/validation.decorator";
 
-export class ForgotPasswordDto {
-  @ApiProperty({
-    type: String,
-    description: "User's email address",
-    example: "test@test.com",
-  })
-  @IsEmail()
-  email: string;
-}
-
-export class ResetPasswordDto {
+export class UpdatePasswordDto {
   @ApiProperty({
     type: String,
     description: "Old password",
@@ -31,7 +21,7 @@ export class ResetPasswordDto {
   @IsNotEmpty()
   @Matches(/^(?=.*\d)(?=.*[!@#$%^&*.?_-])(?=.*[a-z])(?=.*[A-Z]).{8,32}$/, {
     message:
-      "Password must be 6 to 15 characters and must contain a letter, a number, a symbol, one upper case and " +
+      "Password must be 8 to 32 characters and must contain a letter, a number, a symbol, one upper case and " +
       "lower case character.",
   })
   newPassword: string;
