+++ src/api/user/interfaces/users.controller.interface.ts
@@ -1,5 +1,5 @@
 import { IResponseStatus } from "../../../common/interfaces/ResponseStatus.interface";
-import { ForgotPasswordDto, ResetPasswordDto } from "../dtos/password-reset.dto";
+import { ForgotPasswordDto, ResetPasswordDto } from "../dtos/reset-password.dto";
 import { UpdateUserDto } from "../dtos/update-user.dto";
 import { User } from "../entities/user.entity";
